requirements.txt
pandas
flask
pymongo
google-generativeai
python-dotenv
pyarrow
fastparquet
openpyxl
lxml
